rowingwebsite2/requirements.txt
asgiref==3.4.1
awesome-slugify==1.6.5
boto3==1.17.109       
botocore==1.20.109    
Django==3.2.5
django-admin-jqueryui112==1.12.1b5
django-admin-sortable2==1.0
django-appconf==1.0.4
django-cleanup==5.2.0
django-content-gallery==1.0.0b1
django-image-cropping==1.5.0
django-imagekit==4.0.2
django-imagekit-cropper==1.17
django-modeladmin-reorder==0.3.1
django-photologue==3.13
django-sortedm2m==3.0.2
django-sortedmanytomany==0.1.3
django-starcross-gallery==1.0.11
django-storages==1.11.1
django-unused-media==0.2.2
easy-thumbnails==2.7.1
ExifRead==2.3.2
jmespath==0.10.0
pilkit==2.0
Pillow==8.3.1
python-dateutil==2.8.1
python-dotenv==0.18.0
python-magic==0.4.24
pytz==2021.1
regex==2021.7.6
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
Unidecode==0.4.21
urllib3==1.26.6